refactor: components 하위 폴더명 lowerCase로 변경
